+++ src/AngleSharp.Js/Proxies/DomNodeInstance.cs
@@ -10,18 +10,20 @@ namespace AngleSharp.Js
     sealed class DomNodeInstance : ObjectInstance, IDomProxy
     {
         private readonly EngineInstance _instance;
-        private readonly Object _value;
+        private readonly Type _type;
+        private Object _value;
 
         private Dictionary<DomEventDefinition, DomEventDefinition.Registration> _eventHandlers;
         private DomPrototypeState _state;
 
-        public DomNodeInstance(EngineInstance engine, Object value)
+        public DomNodeInstance(EngineInstance engine, Object value, Type type)
             : base(engine.Jint)
         {
             _instance = engine;
             _value = value;
+            _type = type;
 
-            var prototype = engine.GetDomPrototype(value.GetType());
+            var prototype = engine.GetDomPrototype(type);
             Prototype = prototype;
             _state = DomPrototypeState.Of(prototype);
         }
@@ -55,7 +57,11 @@ namespace AngleSharp.Js
 
         public Object Value => _value;
 
+        public Type DomType => _type;
+
         public EngineInstance Instance => _instance;
+
+        public void Update(Object value) => _value = value;
 
         public override object ToObject() => _value;